setup env for restx
